docker_analyse/requirements.txt
flask
flask_cors
flask-restful
redis
google-api-python-client
google-auth-httplib2
google-auth
